+++ jest.config.js
@@ -1,4 +1,6 @@
 module.exports = {
-  testMatch: ['**/*.spec.ts'],
-  moduleDirectories: ['<rootDir>', 'node_modules'],
+  testMatch: ["**/*.spec.ts"],
+  testPathIgnorePatterns: ["<rootDir>/src/__tests__/api"],
+  moduleDirectories: ["<rootDir>", "node_modules"],
+  testTimeout: 30000,
 };