+++ src/components/Directory/Directory.styles.ts
@@ -33,9 +33,17 @@ export const Mobile = styled.div`
 export const Search = styled.div`
   display: flex;
   flex-wrap: wrap;
-  align-items: flex-end;
+  align-items: center;
   flex: 0 1 auto;
   margin-bottom: 24px;
+`;
+
+export const SubSearch = styled.div`
+  display: flex;
+  flex-wrap: wrap;
+  align-items: flex-start;
+  justify-content: space-between;
+  gap: 16px;
 `;
 
 export const Psychologists = styled.div`