+++ src/pages/medecins.tsx
@@ -1,9 +1,11 @@
-import { Container } from "@dataesr/react-dsfr";
+import { Button, Container } from "@dataesr/react-dsfr";
+import { useRouter } from "next/router";
 import React from "react";
 
 import HeadTag from "../components/HeadTag";
 
 const Page = () => {
+  const router = useRouter();
   return (
     <>
       <HeadTag
@@ -251,20 +253,27 @@ const Page = () => {
               </p>
             </div>
           </div>
-        </Container>
-      </section>
-      <section>
-        <div className="fr-container fr-callout fr-callout--pink-tuile fr-my-6w">
-          <div className="fr-grid-row fr-grid-row--gutters fr-grid-row--middle">
-            <div className="fr-col-12">
-              <p className="fr-text--lead">
-                Les coordonnées des psychologues partenaires, conventionnés avec
-                l’Assurance Maladie, seront disponibles à partir du printemps
-                2022.
-              </p>
+          {process.env.NEXT_PUBLIC_NEW_FEATURES === "true" ? (
+            <Button
+              icon="fr-fi-search-line"
+              onClick={() => router.push("/annuaire")}
+            >
+              Trouver un psychologue partenaire
+            </Button>
+          ) : (
+            <div className="fr-container fr-callout fr-callout--pink-tuile fr-my-6w">
+              <div className="fr-grid-row fr-grid-row--gutters fr-grid-row--middle">
+                <div className="fr-col-12">
+                  <p className="fr-text--lead">
+                    Les coordonnées des psychologues partenaires, conventionnés
+                    avec l’Assurance Maladie, seront disponibles à partir du
+                    printemps 2022.
+                  </p>
+                </div>
+              </div>
             </div>
-          </div>
-        </div>
+          )}
+        </Container>
       </section>
       <section>
         <div className="fr-container fr-my-6w">
@@ -303,7 +312,7 @@ const Page = () => {
               <p className="fr-text--lead">
                 J’ai des questions&nbsp;?{" "}
                 <a
-                  href="/faq#medecin"
+                  href="/faq?tab=medecin"
                   className="fr-link fr-fi-question-line fr-link--icon-left"
                 >
                   Je consulte la FAQ