+++ src/services/psychologists.ts
@@ -1,27 +1,61 @@
-import Sequelize from "sequelize";
+import Sequelize, { Op } from "sequelize";
 
 import { models } from "../db/models";
 import { SRID } from "../types/const/geometry";
 import { FILTER } from "../types/enums/filters";
+import { PUBLIC } from "../types/enums/public";
 import { Psychologist } from "../types/psychologist";
+import getAddressCoordinates from "./getAddressCoordinates";
 
-export const getOne = async (id: string) => {
+export const getOne = async (id: string): Promise<Psychologist> => {
+  // @ts-ignore
   return models.Psychologist.findOne({
     raw: true,
-    where: { id },
+    where: { archived: false, id },
   });
 };
 
-export const countAll = async () => models.Psychologist.count();
+export const getByInstructor = async (
+  group: string
+): Promise<Psychologist[]> => {
+  // @ts-ignore
+  return models.Psychologist.findAll({
+    raw: true,
+    where: { archived: false, instructorId: group, state: "accepte" },
+  });
+};
 
+export const countAll = async () =>
+  models.Psychologist.count({
+    where: { archived: false, state: "accepte", visible: true },
+  });
+
+const DEFAULT_PAGE_SIZE = 50;
 export const getAll = async (filters: {
   [key in FILTER]?: string | string[];
 }): Promise<Psychologist[]> => {
+  const pageSize = filters[FILTER.PAGE_SIZE]
+    ? parseInt(filters[FILTER.PAGE_SIZE] as string, 10)
+    : DEFAULT_PAGE_SIZE;
+
+  const where: any = { archived: false, state: "accepte", visible: true };
   const query: Sequelize.FindOptions<any> = {
-    limit: 10,
-    offset: parseInt(filters[FILTER.PAGE_INDEX] as string, 10) * 10,
+    limit: pageSize,
+    offset: parseInt(filters[FILTER.PAGE_INDEX] as string, 10) * pageSize,
     raw: true,
+    where,
   };
+
+  if (filters[FILTER.TELECONSULTATION]) {
+    where.teleconsultation = true;
+  }
+
+  if (filters[FILTER.PUBLIC]) {
+    where.public = {
+      [Op.or]: [filters[FILTER.PUBLIC], PUBLIC.BOTH],
+    };
+  }
+
   if (filters[FILTER.LONGITUDE] && filters[FILTER.LATITUDE]) {
     query.attributes = {
       include: [
@@ -55,12 +89,43 @@ export const saveMany = async (psychologists: Psychologist[]) => {
   return models.Psychologist.bulkCreate(psychologists);
 };
 
+export const update = async (
+  id: string,
+  psychologist: Partial<Psychologist>
+) => {
+  const coordinates = await getAddressCoordinates(psychologist.address);
+  return models.Psychologist.update(
+    {
+      address: psychologist.address,
+      cdsmsp: psychologist.cdsmsp,
+      coordinates: coordinates
+        ? {
+            coordinates: [coordinates.longitude, coordinates.latitude],
+            type: "POINT",
+          }
+        : null,
+      displayEmail: psychologist.displayEmail,
+      email: psychologist.email,
+      firstName: psychologist.firstName,
+      languages: psychologist.languages,
+      lastName: psychologist.lastName,
+      phone: psychologist.phone,
+      public: psychologist.public,
+      teleconsultation: psychologist.teleconsultation,
+      visible: psychologist.visible,
+      website: psychologist.website,
+    },
+    { where: { id } }
+  );
+};
+
 export const updateState = async (newStates: Partial<Psychologist>[]) => {
   return Promise.all(
     newStates.map((newState) =>
       models.Psychologist.update(
         {
           archived: newState.archived,
+          state: newState.state,
         },
         { where: { id: newState.id } }
       )